+++ modules/Attendance/includes/DailySummary.php
@@ -174,7 +174,7 @@ if ( $_REQUEST['search_modfunc']
 }
 
 // @since 9.2.1 SQL use REPLACE() instead of to_char() for MySQL compatibility
-$cal_RET = DBGet( "SELECT DISTINCT SCHOOL_DATE,CONCAT('_', REPLACE(CAST(SCHOOL_DATE AS varchar(10)),'-','')) AS SHORT_DATE
+$cal_RET = DBGet( "SELECT DISTINCT SCHOOL_DATE,CONCAT('_', REPLACE(CAST(SCHOOL_DATE AS char(10)),'-','')) AS SHORT_DATE
 	FROM attendance_calendar
 	WHERE SCHOOL_ID='" . UserSchool() . "'
 	AND SCHOOL_DATE BETWEEN '" . $start_date . "' AND '" . $end_date . "'
@@ -327,7 +327,7 @@ else
 	{
 		// @since 9.2.1 SQL use REPLACE() instead of to_char() for MySQL compatibility
 		$att_sql = "SELECT ad.STATE_VALUE AS STATE_CODE,
-			SCHOOL_DATE,CONCAT('_', REPLACE(CAST(ad.SCHOOL_DATE AS varchar(10)),'-','')) AS SHORT_DATE
+			SCHOOL_DATE,CONCAT('_', REPLACE(CAST(ad.SCHOOL_DATE AS char(10)),'-','')) AS SHORT_DATE
 		FROM attendance_day ad,student_enrollment ssm
 		WHERE ad.STUDENT_ID=ssm.STUDENT_ID
 		AND (('" . DBDate() . "' BETWEEN ssm.START_DATE AND ssm.END_DATE OR ssm.END_DATE IS NULL)
@@ -339,7 +339,7 @@ else
 	else
 	{
 		// @since 9.2.1 SQL use REPLACE() instead of to_char() for MySQL compatibility
-		$att_sql = "SELECT ap.ATTENDANCE_CODE,ap.SCHOOL_DATE,CONCAT('_', REPLACE(CAST(ap.SCHOOL_DATE AS varchar(10)),'-','')) AS SHORT_DATE
+		$att_sql = "SELECT ap.ATTENDANCE_CODE,ap.SCHOOL_DATE,CONCAT('_', REPLACE(CAST(ap.SCHOOL_DATE AS char(10)),'-','')) AS SHORT_DATE
 		FROM attendance_period ap,student_enrollment ssm
 		WHERE ap.STUDENT_ID=ssm.STUDENT_ID
 		AND ap.SCHOOL_DATE BETWEEN '" . $start_date . "' AND '" . $end_date . "'