+++ modules/Food_Service/Students/Reminders.php
@@ -152,11 +152,11 @@ if ( $_REQUEST['modfunc'] === 'save' )
 				AND SYEAR='" . UserSyear() . "'
 				AND (START_DATE<=CURRENT_DATE AND (END_DATE IS NULL OR CURRENT_DATE<=END_DATE)))" );
 
-			// @since 9.3 SQL use CAST(X AS varchar(X)) instead of to_char() for MySQL compatibility
+			// @since 9.3 SQL use CAST(X AS char(X)) instead of to_char() for MySQL compatibility
 			$last_deposit = DBGet( "SELECT (SELECT sum(AMOUNT)
 				FROM food_service_transaction_items
 				WHERE TRANSACTION_ID=fst.TRANSACTION_ID) AS AMOUNT,
-				CAST(fst.TIMESTAMP AS varchar(10)) AS DATE
+				CAST(fst.TIMESTAMP AS char(10)) AS DATE
 			FROM food_service_transactions fst
 			WHERE fst.SHORT_NAME='DEPOSIT'
 			AND fst.ACCOUNT_ID='" . (int) $student['ACCOUNT_ID'] . "'